+++ changelog.md
@@ -9,6 +9,7 @@
 * [changed] Order media files in sequence together
 * [added] Download media files on background
 * [added] Select all media files on the page, or in the selection
+* [added] Page title in media files upload
 
 ## 10-2024